refactor(lens): migrate governance tools to free functions (epic #1655 PR 7/9)
5 tools: get_compliance_status, get_governance_kpis, get_framework_coverage,
get_governance_narrative, get_recent_governance_events. Nested _status
helper kept inline (closure over 4 method-local vars); _grade lifted to
module level (pure). Same pattern as PR 1-6.
