+++ apps/api/app/modules/glens/executor.py
@@ -404,242 +404,7 @@ class Executor:
 
     # ── Compliance ────────────────────────────────────────────────────────────
 
-    def _tool_get_compliance_status(self):
-        """Replicates /guard/verify/evidence DB logic without an HTTP round-trip."""
-        from datetime import timedelta
-        import hashlib
-
-        ws_uuid = uuid.UUID(self.workspace_id)
-        org_ws = _org_ws_subquery(self.db, self.workspace_id)
-        now = datetime.now(timezone.utc)
-        since = now - timedelta(hours=24)
-
-        gc = self.db.query(GuardConfig).filter(GuardConfig.workspace_id == ws_uuid).first()
-        signing_key = self.db.query(WorkspaceSigningKey).filter(
-            WorkspaceSigningKey.workspace_id == ws_uuid
-        ).first()
-
-        policy_count = (
-            self.db.query(WorkspaceCustomRule)
-            .filter(WorkspaceCustomRule.workspace_id.in_(org_ws))
-            .count()
-            + self.db.query(WorkspaceSkillPack)
-            .filter(WorkspaceSkillPack.workspace_id.in_(org_ws))
-            .count()
-        )
-
-        events_24h = (
-            self.db.query(GuardAuditEvent)
-            .filter(
-                GuardAuditEvent.workspace_id == ws_uuid,
-                GuardAuditEvent.ts >= since,
-            )
-            .count()
-        )
-
-        blocked_24h = (
-            self.db.query(GuardAuditEvent)
-            .filter(
-                GuardAuditEvent.workspace_id == ws_uuid,
-                GuardAuditEvent.ts >= since,
-                GuardAuditEvent.decision == "blocked",
-            )
-            .count()
-        )
-
-        # Score computation (mirrors verify.py)
-        score = 0
-        guard_active = bool(gc and gc.enforcement_mode in ("block", "warn", "audit"))
-
-        if gc:
-            if gc.enforcement_mode == "block":
-                score += 30
-            elif gc.enforcement_mode == "warn":
-                score += 15
-            elif gc.enforcement_mode == "audit":
-                score += 5
-            if gc.fail_mode == "fail_closed":
-                score += 15
-            if gc.deny_on_error:
-                score += 10
-
-        if signing_key:
-            score += 15
-
-        if policy_count >= 5:
-            score += 20
-        elif policy_count > 0:
-            score += 10
-
-        if events_24h > 0:
-            score += 10
-
-        fail_closed = gc and gc.fail_mode == "fail_closed"
-
-        _CONTROLS = [
-            ("ASI-01", "Prompt Injection",        "PreToolUse hook intercepts before LLM call"),
-            ("ASI-02", "Insecure Tool Use",       "Guard proxy enforces tool-use policies"),
-            ("ASI-03", "Excessive Agency",        "Turn budgets + max_cost_usd caps"),
-            ("ASI-04", "Unauthorized Escalation", "RBAC + require_permission() on all endpoints"),
-            ("ASI-05", "Trust Boundary Violation","All LLM traffic routed through Guard proxy"),
-            ("ASI-06", "Insufficient Logging",    "guard_audit_events with SHA-256 hash chain"),
-            ("ASI-07", "Insecure Identity",       "agent_role_id + member tokens per agent"),
-            ("ASI-08", "Policy Bypass",           "fail_mode=fail_closed blocks on Guard outage"),
-            ("ASI-09", "Supply Chain Integrity",  "Signed policies (signing_key)"),
-            ("ASI-10", "Behavioral Anomaly",      "Session scanning + violations_count tracking"),
-        ]
-
-        def _status(asi: str) -> str:
-            if asi in ("ASI-01", "ASI-02", "ASI-05"):
-                return "active" if guard_active else "missing"
-            if asi in ("ASI-03", "ASI-04"):
-                return "active"
-            if asi == "ASI-06":
-                if guard_active and events_24h > 0:
-                    return "active"
-                return "partial" if guard_active else "missing"
-            if asi == "ASI-07":
-                return "partial"
-            if asi == "ASI-08":
-                return "active" if fail_closed else "partial"
-            if asi == "ASI-09":
-                return "active" if signing_key else "missing"
-            if asi == "ASI-10":
-                return "partial"
-            return "missing"
-
-        controls = [
-            {"control_id": asi, "name": name, "status": _status(asi), "conduct_enforcement": enforcement}
-            for asi, name, enforcement in _CONTROLS
-        ]
-
-        active_count = sum(1 for c in controls if c["status"] == "active")
-        coverage_pct = round(active_count / len(controls) * 100)
-
-        def _grade(s: int) -> str:
-            for threshold, letter in [(85, "A"), (65, "B"), (45, "C"), (25, "D"), (0, "F")]:
-                if s >= threshold:
-                    return letter
-            return "F"
-
-        return {
-            "grade": _grade(score),
-            "score": score,
-            "coverage_pct": coverage_pct,
-            "controls": controls,
-            "events_24h": {"total": events_24h, "blocked": blocked_24h},
-        }
-
     # ── Governance ────────────────────────────────────────────────────────────
-
-    def _tool_get_governance_kpis(self):
-        from datetime import timedelta
-        from sqlalchemy import func as sa_func
-
-        ws_uuid = uuid.UUID(self.workspace_id)
-        org_ws = _org_ws_subquery(self.db, self.workspace_id)
-        now = datetime.now(timezone.utc)
-        today_start = now.replace(hour=0, minute=0, second=0, microsecond=0)
-        month_start = now.replace(day=1, hour=0, minute=0, second=0, microsecond=0)
-
-        base_q = self.db.query(GuardAuditEvent).filter(
-            GuardAuditEvent.workspace_id.in_(org_ws)
-        )
-
-        events_today = base_q.filter(GuardAuditEvent.ts >= today_start).count()
-
-        blocked_today = base_q.filter(
-            GuardAuditEvent.ts >= today_start,
-            GuardAuditEvent.decision == "blocked",
-        ).count()
-
-        active_developers_today = (
-            self.db.query(sa_func.count(sa_func.distinct(GuardAuditEvent.user_email)))
-            .filter(
-                GuardAuditEvent.workspace_id.in_(org_ws),
-                GuardAuditEvent.ts >= today_start,
-                GuardAuditEvent.user_email.isnot(None),
-            )
-            .scalar() or 0
-        )
-
-        blocks_mtd = base_q.filter(
-            GuardAuditEvent.ts >= month_start,
-            GuardAuditEvent.decision == "blocked",
-        ).count()
-
-        # Approximate risk avoided: $0.01 per blocked event (no cost field on blocked events)
-        risk_avoided_usd_mtd = round(blocks_mtd * 0.01, 2)
-
-        return {
-            "events_today": events_today,
-            "blocked_today": blocked_today,
-            "active_developers_today": int(active_developers_today),
-            "blocks_mtd": blocks_mtd,
-            "risk_avoided_usd_mtd": risk_avoided_usd_mtd,
-        }
-
-    def _tool_get_framework_coverage(self):
-        from sqlalchemy import func as sa_func
-
-        org_ws = _org_ws_subquery(self.db, self.workspace_id)
-
-        _PACK_TO_FRAMEWORK = {
-            "conduct-owasp":    "OWASP",
-            "conduct-soc2":     "SOC2",
-            "conduct-hipaa":    "HIPAA",
-            "conduct-pci-dss":  "PCI_DSS",
-            "conduct-base":     "Base",
-            "conduct-iso42001": "ISO_42001",
-            "conduct-gdpr":     "GDPR",
-            "conduct-nist":     "NIST",
-            "conduct-nis2":     "NIS2",
-            "conduct-dora":     "DORA",
-        }
-
-        # Installed packs for this workspace (org-scoped)
-        installed_wsps = (
-            self.db.query(WorkspaceSkillPack)
-            .filter(WorkspaceSkillPack.workspace_id.in_(org_ws))
-            .all()
-        )
-
-        # Collect unique pack slugs installed
-        seen_slugs: set[str] = set()
-        wsp_by_slug: dict[str, WorkspaceSkillPack] = {}
-        for wsp in installed_wsps:
-            if wsp.pack_slug not in seen_slugs:
-                seen_slugs.add(wsp.pack_slug)
-                wsp_by_slug[wsp.pack_slug] = wsp
-
-        # Fetch rules counts from SkillPack catalog (latest version per slug)
-        rules_count_by_slug: dict[str, int] = {}
-        if seen_slugs:
-            catalog_rows = (
-                self.db.query(SkillPack.slug, SkillPack.rules)
-                .filter(SkillPack.slug.in_(list(seen_slugs)))
-                .all()
-            )
-            # Sum rules across all versions (or just take any — rules are per slug+version)
-            for slug, rules in catalog_rows:
-                cnt = len(rules) if isinstance(rules, list) else 0
-                rules_count_by_slug[slug] = max(rules_count_by_slug.get(slug, 0), cnt)
-
-        frameworks = [
-            {
-                "pack_slug": slug,
-                "framework": _PACK_TO_FRAMEWORK.get(slug, slug),
-                "rules_count": rules_count_by_slug.get(slug, 0),
-                "installed_at": wsp_by_slug[slug].installed_at.isoformat()
-                    if wsp_by_slug[slug].installed_at else None,
-            }
-            for slug in sorted(seen_slugs)
-        ]
-
-        return {
-            "installed_count": len(frameworks),
-            "frameworks": frameworks,
-        }
 
     def _tool_get_correlated_events(self, decision: str = "blocked", since: str | None = None, until: str | None = None, limit: int = 50):
         from app.modules.guard.models import GuardSession as _GS
@@ -690,45 +455,6 @@ class Executor:
             for m in result.by_member
         ]
         return {"total_tokens_saved": total, "total_cost_saved_usd": round(usd, 4), "by_member": members}
-
-    def _tool_get_governance_narrative(self):
-        from app.routers.governance import get_narrative as _get_narrative, _compute_framework_coverage
-        try:
-            result = _get_narrative(db=self.db, workspace_id=self.workspace_id)
-            return {"paragraph": result.paragraph, "source": result.source}
-        except Exception as e:
-            return {"paragraph": f"Narrative unavailable: {e}", "source": "error"}
-
-    def _tool_get_recent_governance_events(self, limit: int = 15, decision: str | None = None, since: str | None = None, until: str | None = None):
-        if decision:
-            decision = {"block": "blocked", "warn": "warned", "audit": "audited", "allow": "allowed"}.get(decision, decision)
-        # when a full date range is given and no explicit limit, fetch all events in that window
-        if since and until and limit == 15:
-            limit = 500
-        org_ws = _org_ws_subquery(self.db, self.workspace_id)
-        q = self.db.query(GuardAuditEvent).filter(GuardAuditEvent.workspace_id.in_(org_ws))
-        if decision:
-            q = q.filter(GuardAuditEvent.decision == decision)
-        if since:
-            q = q.filter(GuardAuditEvent.ts >= since)
-        if until:
-            q = q.filter(GuardAuditEvent.ts <= until)
-        rows = (
-            q.order_by(GuardAuditEvent.ts.desc())
-            .limit(min(limit, 100))
-            .all()
-        )
-        return [
-            {
-                "id": str(e.id),
-                "ts": e.ts.isoformat(),
-                "decision": e.decision,
-                "rule_id": e.rule_id,
-                "ai_tool": e.ai_tool,
-                "user_email": e.user_email,
-            }
-            for e in rows
-        ]
 
     # ── Workflow inventory + block roll-up ──────────────────────────────────
